Reimplemented MarkdownParser.java
A full reimplementation of MarkdownParser class.

Also, couple files added to .gitignore after running maven.

I've decided that the previous implementation was hard to work with and not very extendable for the future.
Writing the code took about 1:40h, and then another 2h for writing the documentation.

I've kept the class core exposed structure the same, so there is no need to update other parts of the codebase.

There is a full basic documentation in the code (including on how to extend it).

This implementation focuses on extensibility and maintenance (clean structure and no overengineering).